refactor: 패키지분리, 네이밍 변경, view 비즈니스 로직 제거
- domain, view 분리
- history -> result 변경
- OutputView 비즈니스 로직 제거
