+++ src/main/java/race/RaceGameApplication.java
@@ -1,5 +1,11 @@
 package race;
 
+import race.domain.RaceGameFactory;
+import race.domain.RaceGameResult;
+import race.domain.RandomRaceGameFactory;
+import race.view.InputView;
+import race.view.OutputView;
+
 public class RaceGameApplication {
 
 	public static void main(String[] args) {
@@ -7,9 +13,9 @@ public class RaceGameApplication {
 			String names = InputView.askNameOfCarToUser();
 			int tryTimesToUser = InputView.askNumberOfTryTimesToUser();
 			RaceGameFactory raceGameFactory = new RandomRaceGameFactory(names, tryTimesToUser);
-			RaceGameResult history = raceGameFactory.start();
-			OutputView.printDistanceStatus(history);
-			OutputView.printRaceGameWinner(history);
+			RaceGameResult gameResult = raceGameFactory.start();
+			OutputView.printDistanceStatus(gameResult);
+			OutputView.printRaceGameWinner(gameResult);
 		} catch (Exception e) {
 			System.out.println(e.getMessage());
 		}
