+++ src/main/java/race/domain/CarGroup.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 import java.util.Arrays;
 import java.util.Collections;
