+++ src/main/java/race/domain/IllegalInputNameException.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 public class IllegalInputNameException extends IllegalArgumentException {
 	public static final String ILLEGAL_INPUT_FORMAT_MESSAGE = "자동차 이름은 %d자를 초과할 수 없습니다.";
