+++ src/main/java/race/domain/MovePolicy.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 public interface MovePolicy {
 	int INCREASE_DISTANCE_NUMBER = 1;
