+++ src/main/java/race/domain/RaceGameFactory.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 public interface RaceGameFactory {
 	RaceGameResult start();
